+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # COFantasy: Historique des changements (depuis la version 1.0)
 ## 3.05
 ### Corrections de bugs
+* Correction de bugs sur le radar mental
 * Correction d'un bug sur le pacifisme
 
 ### Autres améliorations